pyproject.toml
[tool.poetry]
name = "code-dot-org-analytics-infrastructure"
version = "0.1.0"
description = ""
authors = ["Code.org"]

[tool.poetry.dependencies]
python = "^3.9"
redshift-connector = "^2.0.905"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
